requirements.txt
Django 3.1.1
setuptools 50.3.0
psycopg2-binary 2.8.6
django-ckeditor 6.0.0
Pillow 7.2.0